requirements.txt
requests
python-dotenv
python-telegram-bot
aiohttp